@@ -1,27 +1,71 @@
-[tool.black]
-line-length = 88
-target-version = ['py38']
-include = '\.pyi?$'
+[tool.ruff]
+# 目标 Python 版本（项目要求 Python 3.10+）
+target-version = "py310"
 
-[tool.isort]
-profile = "black"
-multi_line_output = 3
-line_length = 88
-include_trailing_comma = true
-force_grid_wrap = 0
-use_parentheses = true
-ensure_newline_before_comments = true
+# 每行最大字符数
+line-length = 100
 
-[tool.mypy]
-python_version = "3.8"
-warn_return_any = false
-warn_unused_configs = true
-disallow_untyped_defs = false
-check_untyped_defs = false
-disallow_any_generics = false
-no_implicit_optional = false
+# 扫描的文件
+include = ["*.py", "*.pyi"]
 
-[tool.pytest]
-testpaths = ["tests"]
-python_files = ["test_*.py", "*_test.py"]
-addopts = "-v --cov=app --cov-report=term-missing"
+# 排除的目录
+exclude = [
+    ".git",
+    "venv",
+    "env",
+    "__pycache__",
+    "*.egg-info",
+    "node_modules",
+    "data",
+    "logs",
+    ".pytest_cache",
+]
+
+[tool.ruff.lint]
+# 启用的规则集
+select = [
+    "E",   # pycodestyle errors
+    "W",   # pycodestyle warnings
+    "F",   # pyflakes
+    "I",   # isort (import 排序)
+    "N",   # pep8-naming
+    "UP",  # pyupgrade (现代 Python 语法)
+    "B",   # flake8-bugbear (常见错误)
+    "C4",  # flake8-comprehensions
+    "SIM", # flake8-simplify
+]
+
+# 忽略的规则
+ignore = [
+    "E501",  # 行过长（由 formatter 处理）
+    "B008",  # 函数调用作为默认参数（Flask 路由常见）
+    "E402",  # 模块导入不在顶部（某些场景需要）
+    "N818",  # 异常名以Error结尾（现有代码设计）
+    "E722",  # 裸 except（现有代码逻辑）
+    "SIM102", "SIM103", "SIM105", "SIM108", "SIM401",  # 代码简化建议（可选优化）
+    "B007",  # 循环变量未使用
+    "B904",  # raise ... from err（可选）
+]
+
+# 允许自动修复的规则
+fixable = ["ALL"]
+unfixable = []
+
+[tool.ruff.lint.per-file-ignores]
+# __init__.py 允许未使用的导入
+"__init__.py" = ["F401"]
+# 测试文件允许更宽松的规则
+"tests/**/*.py" = ["S101"]  # 允许使用 assert
+
+[tool.ruff.format]
+# 使用双引号
+quote-style = "double"
+
+# 缩进样式
+indent-style = "space"
+
+# 保留魔法尾随逗号
+skip-magic-trailing-comma = false
+
+# 行结束符
+line-ending = "auto"
