@@ -2,7 +2,7 @@ import json
 import os
 import re
 from pathlib import Path
-from typing import ClassVar, List, Optional, Union
+from typing import ClassVar, Optional
 
 from flask import current_app
 
@@ -16,18 +16,16 @@ class TextAnonymizer:
             cls._instance = cls()
         return cls._instance
 
-    def __init__(self, sensitive_words_path: Optional[str] = None) -> None:
-        self.sensitive_words: List[str] = []
+    def __init__(self, sensitive_words_path: str | None = None) -> None:
+        self.sensitive_words: list[str] = []
         if sensitive_words_path is None and current_app:
             # 使用配置中的路径
-            sensitive_words_path = current_app.config["FILE_CONFIG"][
-                "SENSITIVE_WORDS_FILE"
-            ]
+            sensitive_words_path = current_app.config["FILE_CONFIG"]["SENSITIVE_WORDS_FILE"]
 
         if sensitive_words_path:
             self.load_sensitive_words(sensitive_words_path)
 
-        self.patterns: List[str] = [
+        self.patterns: list[str] = [
             r"(909|ARJ)/B-?[A-Z0-9]{4}",
             r"B-?[A-Z0-9]{4}",
             r"10\d{3}",
@@ -35,7 +33,7 @@ class TextAnonymizer:
             r"[A-Z]{2}\d{4}",
         ]
 
-    def load_sensitive_words(self, file_path: Union[str, Path]) -> None:
+    def load_sensitive_words(self, file_path: str | Path) -> None:
         try:
             # 确保文件路径是 Path 对象
             file_path = Path(file_path)
@@ -60,7 +58,7 @@ class TextAnonymizer:
         except Exception as e:
             print(f"读取敏感词文件时出错：{str(e)}")
 
-    def add_sensitive_words(self, new_words: Union[str, List[str]]) -> None:
+    def add_sensitive_words(self, new_words: str | list[str]) -> None:
         if isinstance(new_words, str):
             new_words = [new_words]
         for word in new_words:
@@ -69,7 +67,7 @@ class TextAnonymizer:
         # 重新排序
         self.sensitive_words.sort(key=len, reverse=True)
 
-    def add_patterns(self, new_patterns: Union[str, List[str]]) -> None:
+    def add_patterns(self, new_patterns: str | list[str]) -> None:
         if isinstance(new_patterns, str):
             new_patterns = [new_patterns]
         for pattern in new_patterns:
@@ -87,10 +85,10 @@ class TextAnonymizer:
             return text
         return text
 
-    def get_sensitive_words(self) -> List[str]:
+    def get_sensitive_words(self) -> list[str]:
         return self.sensitive_words
 
-    def get_patterns(self) -> List[str]:
+    def get_patterns(self) -> list[str]:
         return self.patterns
 
 
@@ -104,17 +102,17 @@ def anonymize_text(text: str) -> str:
     return get_anonymizer().anonymize_text(text)
 
 
-def add_sensitive_words(words: Union[str, List[str]]) -> None:
+def add_sensitive_words(words: str | list[str]) -> None:
     return get_anonymizer().add_sensitive_words(words)
 
 
-def add_patterns(patterns: Union[str, List[str]]) -> None:
+def add_patterns(patterns: str | list[str]) -> None:
     return get_anonymizer().add_patterns(patterns)
 
 
-def get_sensitive_words() -> List[str]:
+def get_sensitive_words() -> list[str]:
     return get_anonymizer().get_sensitive_words()
 
 
-def get_patterns() -> List[str]:
+def get_patterns() -> list[str]:
     return get_anonymizer().get_patterns()
